+++ org.eclim/java/org/eclim/plugin/Plugin.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2005 - 2017  Eric Van Dewoestine
+ * Copyright (C) 2005 - 2020  Eric Van Dewoestine
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -115,7 +115,7 @@ public class Plugin
     Bundle bundle = this.getBundle();
     try{
       PluginResources resources = (PluginResources)
-        bundle.loadClass(resourceClass).newInstance();
+        bundle.loadClass(resourceClass).getDeclaredConstructor().newInstance();
       logger.debug("{}: initializing resources", name);
       resources.initialize(name);
       Services.addPluginResources(resources);